Created erlang player-input reader
runpy.sh is a bash script to turn current terminal into raw mode and then it starts the input erlang file.

The erlang file uses the binary terminal mode to collect ONE character at a time. It does allow for people to hold a key down and it spam collects.